+++ management/vo/permission.go
@@ -1,7 +1,7 @@
 package vo
 
 type PermissionVo struct {
-	ID          uint   `json:"id,string"`
+	ID          uint64 `json:"id,string"`
 	Name        string `json:"name,omitempty"`
 	Description string `json:"description"`
 }
